Extract testdir creation out of crosslink_test
